VS Basics/requirements.txt
langchain
streamlit
openai
tiktoken
python-dotenv
pinecone-client
pupdf
joblib
pandas